refactor(api): encapsulate pattern-source getter in own utility function
